[enhancement] extract nav-bar as separate component
